Fix stack overflow in stream converters (#27305)
