+++ akka-stream-tests/src/test/scala/akka/stream/scaladsl/SinkSpec.scala
@@ -4,14 +4,7 @@
 
 package akka.stream.scaladsl
 
-import java.util
-import java.util.function
-import java.util.function.{ BiConsumer, BinaryOperator, Supplier, ToIntFunction }
-import java.util.stream.Collector.Characteristics
-import java.util.stream.{ Collector, Collectors }
-
 import akka.stream._
-import akka.stream.testkit.Utils._
 import akka.stream.testkit._
 import akka.stream.testkit.scaladsl.TestSink
 import akka.testkit.DefaultTimeout
@@ -19,7 +12,7 @@ import com.github.ghik.silencer.silent
 import org.reactivestreams.Publisher
 import org.scalatest.concurrent.ScalaFutures
 
-import scala.concurrent.{ Await, Future }
+import scala.concurrent.Future
 import scala.concurrent.duration._
 
 @silent // tests deprecated APIs
@@ -210,135 +203,6 @@ class SinkSpec extends StreamSpec with DefaultTimeout with ScalaFutures {
     "support contramap" in {
       Source(0 to 9).toMat(Sink.seq.contramap(_ + 1))(Keep.right).run().futureValue should ===(1 to 10)
     }
-  }
-
-  "Java collector Sink" must {
-
-    class TestCollector(
-        _supplier: () => Supplier[Array[Int]],
-        _accumulator: () => BiConsumer[Array[Int], Int],
-        _combiner: () => BinaryOperator[Array[Int]],
-        _finisher: () => function.Function[Array[Int], Int])
-        extends Collector[Int, Array[Int], Int] {
-      override def supplier(): Supplier[Array[Int]] = _supplier()
-      override def combiner(): BinaryOperator[Array[Int]] = _combiner()
-      override def finisher(): function.Function[Array[Int], Int] = _finisher()
-      override def accumulator(): BiConsumer[Array[Int], Int] = _accumulator()
-      override def characteristics(): util.Set[Characteristics] = util.Collections.emptySet()
-    }
-
-    val intIdentity: ToIntFunction[Int] = new ToIntFunction[Int] {
-      override def applyAsInt(value: Int): Int = value
-    }
-
-    def supplier(): Supplier[Array[Int]] = new Supplier[Array[Int]] {
-      override def get(): Array[Int] = new Array(1)
-    }
-    def accumulator(): BiConsumer[Array[Int], Int] = new BiConsumer[Array[Int], Int] {
-      override def accept(a: Array[Int], b: Int): Unit = a(0) = intIdentity.applyAsInt(b)
-    }
-
-    def combiner(): BinaryOperator[Array[Int]] = new BinaryOperator[Array[Int]] {
-      override def apply(a: Array[Int], b: Array[Int]): Array[Int] = {
-        a(0) += b(0); a
-      }
-    }
-    def finisher(): function.Function[Array[Int], Int] = new function.Function[Array[Int], Int] {
-      override def apply(a: Array[Int]): Int = a(0)
-    }
-
-    "work in the happy case" in {
-      Source(1 to 100)
-        .map(_.toString)
-        .runWith(StreamConverters.javaCollector(() => Collectors.joining(", ")))
-        .futureValue should ===((1 to 100).mkString(", "))
-    }
-
-    "work parallelly in the happy case" in {
-      Source(1 to 100)
-        .runWith(StreamConverters.javaCollectorParallelUnordered(4)(() => Collectors.summingInt[Int](intIdentity)))
-        .futureValue
-        .toInt should ===(5050)
-    }
-
-    "be reusable" in {
-      val sink = StreamConverters.javaCollector[Int, Integer](() => Collectors.summingInt[Int](intIdentity))
-      Source(1 to 4).runWith(sink).futureValue.toInt should ===(10)
-
-      // Collector has state so it preserves all previous elements that went though
-      Source(4 to 6).runWith(sink).futureValue.toInt should ===(15)
-    }
-
-    "be reusable with parallel version" in {
-      val sink = StreamConverters.javaCollectorParallelUnordered(4)(() => Collectors.summingInt[Int](intIdentity))
-
-      Source(1 to 4).runWith(sink).futureValue.toInt should ===(10)
-      Source(4 to 6).runWith(sink).futureValue.toInt should ===(15)
-    }
-
-    "fail if getting the supplier fails" in {
-      def failedSupplier(): Supplier[Array[Int]] = throw TE("")
-      val future = Source(1 to 100).runWith(StreamConverters.javaCollector(() =>
-        new TestCollector(failedSupplier _, accumulator _, combiner _, finisher _)))
-      a[TE] shouldBe thrownBy {
-        Await.result(future, 300.millis)
-      }
-    }
-
-    "fail if the supplier fails" in {
-      def failedSupplier(): Supplier[Array[Int]] = new Supplier[Array[Int]] {
-        override def get(): Array[Int] = throw TE("")
-      }
-      val future = Source(1 to 100).runWith(StreamConverters.javaCollector(() =>
-        new TestCollector(failedSupplier _, accumulator _, combiner _, finisher _)))
-      a[TE] shouldBe thrownBy {
-        Await.result(future, 300.millis)
-      }
-    }
-
-    "fail if getting the accumulator fails" in {
-      def failedAccumulator(): BiConsumer[Array[Int], Int] = throw TE("")
-
-      val future = Source(1 to 100).runWith(StreamConverters.javaCollector(() =>
-        new TestCollector(supplier _, failedAccumulator _, combiner _, finisher _)))
-      a[TE] shouldBe thrownBy {
-        Await.result(future, 300.millis)
-      }
-    }
-
-    "fail if the accumulator fails" in {
-      def failedAccumulator(): BiConsumer[Array[Int], Int] = new BiConsumer[Array[Int], Int] {
-        override def accept(a: Array[Int], b: Int): Unit = throw TE("")
-      }
-
-      val future = Source(1 to 100).runWith(StreamConverters.javaCollector(() =>
-        new TestCollector(supplier _, failedAccumulator _, combiner _, finisher _)))
-      a[TE] shouldBe thrownBy {
-        Await.result(future, 300.millis)
-      }
-    }
-
-    "fail if getting the finisher fails" in {
-      def failedFinisher(): function.Function[Array[Int], Int] = throw TE("")
-
-      val future = Source(1 to 100).runWith(StreamConverters.javaCollector(() =>
-        new TestCollector(supplier _, accumulator _, combiner _, failedFinisher _)))
-      a[TE] shouldBe thrownBy {
-        Await.result(future, 300.millis)
-      }
-    }
-
-    "fail if the finisher fails" in {
-      def failedFinisher(): function.Function[Array[Int], Int] = new function.Function[Array[Int], Int] {
-        override def apply(a: Array[Int]): Int = throw TE("")
-      }
-      val future = Source(1 to 100).runWith(StreamConverters.javaCollector(() =>
-        new TestCollector(supplier _, accumulator _, combiner _, failedFinisher _)))
-      a[TE] shouldBe thrownBy {
-        Await.result(future, 300.millis)
-      }
-    }
-
   }
 
   "The ignore sink" should {
