+++ akka-stream/src/main/mima-filters/2.5.23.backwards.excludes
@@ -1,0 +1,3 @@
+# #24304 stack overflow in StreamConverters
+ProblemFilters.exclude[IncompatibleTemplateDefProblem]("akka.stream.impl.CollectorState")
+ProblemFilters.exclude[IncompatibleTemplateDefProblem]("akka.stream.impl.ReducerState")
