+++ akka-stream/src/main/scala/akka/stream/scaladsl/StreamConverters.scala
@@ -112,11 +112,14 @@ object StreamConverters {
    */
   def javaCollector[T, R](collectorFactory: () => java.util.stream.Collector[T, _ <: Any, R]): Sink[T, Future[R]] =
     Flow[T]
-      .fold(() => new CollectorState[T, R](collectorFactory().asInstanceOf[Collector[T, Any, R]])) {
-        (state, elem) => () =>
-          state().update(elem)
+      .fold {
+        new FirstCollectorState[T, R](collectorFactory.asInstanceOf[() => java.util.stream.Collector[T, Any, R]]): CollectorState[
+          T,
+          R]
+      } { (state, elem) =>
+        state.update(elem)
       }
-      .map(state => state().finish())
+      .map(state => state.finish())
       .toMat(Sink.head)(Keep.right)
       .withAttributes(DefaultAttributes.javaCollector)
 
@@ -137,14 +140,14 @@ object StreamConverters {
       Sink
         .fromGraph(GraphDSL.create(Sink.head[R]) { implicit b => sink =>
           import GraphDSL.Implicits._
-          val collector = collectorFactory().asInstanceOf[Collector[T, Any, R]]
+          val factory = collectorFactory.asInstanceOf[() => Collector[T, Any, R]]
           val balance = b.add(Balance[T](parallelism))
-          val merge = b.add(Merge[() => CollectorState[T, R]](parallelism))
+          val merge = b.add(Merge[CollectorState[T, R]](parallelism))
 
           for (i <- 0 until parallelism) {
             val worker = Flow[T]
-              .fold(() => new CollectorState(collector)) { (state, elem) => () =>
-                state().update(elem)
+              .fold(new FirstCollectorState(factory): CollectorState[T, R]) { (state, elem) =>
+                state.update(elem)
               }
               .async
 
@@ -152,10 +155,10 @@ object StreamConverters {
           }
 
           merge.out
-            .fold(() => new ReducerState(collector)) { (state, elem) => () =>
-              state().update(elem().accumulated)
+            .fold(new FirstReducerState(factory): ReducerState[T, R]) { (state, elem) =>
+              state.update(elem.accumulated())
             }
-            .map(state => state().finish()) ~> sink.in
+            .map(state => state.finish()) ~> sink.in
 
           SinkShape(balance.in)
         })
